--- Web_App/Models/ProductModel.cs
@@ -1,11 +1,0 @@
-﻿namespace Web_App.Models
-{
-    public class ProductModel
-    { 
-            public int Id { get; set; }
-            public string Name { get; set; }
-            public string Description { get; set; }
-            public string ImageUrl { get; set; }
-       
-    }
-}
