--- Web_App/Models/RefreshToken.cs
@@ -1,7 +1,0 @@
-﻿namespace Web_App.Models
-{
-    public class RefreshToken
-    {
-        public string Token { get; set; }
-    }
-}
