+++ Web_App/Rest/Authorization/Models/AuthorizationModel.cs
@@ -1,6 +1,6 @@
-﻿namespace Web_App.Models
+﻿namespace Web_App.Rest.Authorization.Models
 {
-    public class UserLogin
+    public class AuthorizationModel
     {
         public string Login { get; set; }
         public string Password { get; set; }
