+++ Web_App/Rest/Authorization/Models/RegisterModel.cs
@@ -1,6 +1,6 @@
-﻿namespace Web_App.Models
+﻿namespace Web_App.Rest.Authorization.Models
 {
-    public class SignUpModel
+    public class RegisterModel
     {
         public string Login { get; set; }
         public string Email { get; set; }
