+++ Web_App/Rest/JWT/Model/RefreshTokenModel.cs
@@ -1,0 +1,7 @@
+﻿namespace Web_App.Rest.JWT.Model
+{
+    public class RefreshTokenModel
+    {
+        public string Token { get; set; }
+    }
+}
