+++ Web_App/Rest/JWT/Services/ITokenService.cs
@@ -1,7 +1,7 @@
-﻿using Web_App.Models;
+﻿using Web_App.Rest.JWT.Model;
+using Web_App.Rest.User.Models;
 
-
-namespace Web_App.Services
+namespace Web_App.Rest.JWT.Services
 {
     public interface ITokenService
     {
