+++ Web_App/Rest/Product/Model/ProductModel.cs
@@ -1,0 +1,11 @@
+﻿namespace Web_App.Rest.Product.Model
+{
+    public class ProductModel
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+        public string Description { get; set; }
+        public string ImageUrl { get; set; }
+
+    }
+}
